Move stylesheets, logging and connection reset
