+++ arches_project/core/views/logging.py
@@ -1,0 +1,9 @@
+def enable_logging(self):
+    if self.dlg.enableLoggingCheckbox.isChecked():
+        self.dlg.tabWidget.setTabVisible(5, True)
+
+    elif not self.dlg.enableLoggingCheckbox.isChecked():
+        self.dlg.tabWidget.setTabVisible(5, False)
+
+
+#TODO needs expanding into class and functionality
